farmstack-backend/requirements_servvia2.txt
# ServVia 2. 0 Additional Requirements
neo4j==5.14.0
langgraph==0. 0.40
langchain-openai==0.0.5
biopython==1.82
httpx==0.27.0
django-redis==5.4.0
